+++ sdk/src/org.graalvm.launcher/src/org/graalvm/launcher/PolyglotLauncher.java
@@ -50,12 +50,6 @@ public final class PolyglotLauncher extends Launcher {
 
     private String mainLanguage = null;
 
-    // Commented out because of GR-6870
-    // static {
-    // // TODO remove temporary hack to initialize engine if not yet initialized.
-    // Source.create("js", "");
-    // }
-
     @Override
     protected void printHelp(OptionCategory maxCategory) {
         Engine engine = Engine.create();